:sparkles: feat: (front-end) Separando o Formulário de Cadastro do Processo da página principal.
