+++ Desenvolvimento/FrontEnd/sged/src/pages/manipulation/process/addprocess/index.jsx
@@ -1,6 +1,4 @@
 import React, { useEffect, useState } from "react";
-import Select from "react-select";
-import { HouseLine, User } from '@phosphor-icons/react';
 
 // Importa o arquivo CSS
 import './styles.css';
@@ -13,6 +11,7 @@ import ConnectionService from "../../../../object/service/connection";
 import ListModule from "../../../../object/modules/list";
 import SelectModule from '../../../../object/modules/select';
 
+import ProcessForm from './form/processForm';
 import DocumentComponent from './documents/documentComponent';
 
 const AddProcess = () => {
@@ -185,7 +184,7 @@ const AddProcess = () => {
       idTipoDocumentoEtapa: data.typeId || null,
       idResponsavel: data.responsibleId || null,
     })));
-  
+
     // Constrói o objeto de dados do processo
     const dataProcess = {
       IdentificacaoProcesso: process.identificationNumber,
@@ -193,20 +192,20 @@ const AddProcess = () => {
       SituacaoProcesso: process.processSituation || "",
       DataAprovacao: process.approvationDate || "",
       Status: process.processStatus || 0,
-  
+
       IdImovel: selectBox_Realstate.selectedOption.value,
       IdTipoProcesso: selectBox_TypesProcess.selectedOption.value,
       IdEngenheiro: selectBox_Engineer.selectedOption.value || null,
       IdFiscal: selectBox_Supervisor.selectedOption.value || null,
       IdResponsavel: selectBox_UserResponsible.selectedOption.value || null,
       IdAprovador: selectBox_UserApprover.selectedOption.value || null,
-  
+
       DocumentosProcessoDTOs: documentList
     };
-  
+
     await connection.endpoint("Processo").action("PostAllDatas").post(dataProcess);
     list_Engineers.setList(connection.getList());
-  };  
+  };
 
   // Função para converter arquivo em bytes
   const convertFileToBytes = async (file) => {
@@ -251,7 +250,6 @@ const AddProcess = () => {
 
   useEffect(() => {
     if (selectBox_Realstate.selectedOption.value) {
-      setCurrentImageIndex(0);
       GetRealstate(selectBox_Realstate.selectedOption.value);
     }
   }, [selectBox_Realstate.selectedOption]);
@@ -390,23 +388,6 @@ const AddProcess = () => {
   }, [selectBox_Supervisor.selectedOption]);
 
 
-  // Imagem -------------------------------------------------------------------------------------------------------------------------------------------
-
-  const [currentImageIndex, setCurrentImageIndex] = useState(0); // Índice da imagem atual
-
-  useEffect(() => {
-    if (realstate.imagemImovel && realstate.imagemImovel.length > 1) {
-      const interval = setInterval(() => {
-        setCurrentImageIndex((prevIndex) =>
-          prevIndex === realstate.imagemImovel.length - 1 ? 0 : prevIndex + 1
-        );
-      }, 5000); // A cada 5 segundos
-
-      return () => clearInterval(interval);
-    }
-  }, [realstate.imagemImovel]);
-
-
   // Gerenciar a lista de Etapas:
   const [expandedRows, setExpandedRows] = useState([]); // Controla quais etapas estão expandidas
   const [typeDocumentStagesData, setTypeDocumentStagesData] = useState({}); // Armazena os dados de TipoDocumentEtapa
@@ -516,344 +497,32 @@ const AddProcess = () => {
     <LayoutPage>
       <LinkTitle pageName="Cadastrar Processo" otherRoute="Processo" />
       <div className="mt-8">
-        <div className="flex">
-          {/* Imóvel: ----------------------------------------------------------------------------------------------------*/}
+        <ProcessForm
+          realstate={realstate}
+          selectBox_Realstate={selectBox_Realstate}
+          list_Realstate={list_Realstate}
+          owner={owner}
+          taxpayer={taxpayer}
+          use={use}
+          occupation={occupation}
+          selectBox_TypesProcess={selectBox_TypesProcess}
+          list_TypesProcess={list_TypesProcess}
+          typeProcess={typeProcess}
+          process={process}
+          setProcess={setProcess}
+          engineer={engineer}
+          selectBox_Engineer={selectBox_Engineer}
+          list_Users={list_Users}
+          supervisor={supervisor}
+          selectBox_Supervisor={selectBox_Supervisor}
+          userResponsible={userResponsible}
+          selectBox_UserResponsible={selectBox_UserResponsible}
+          userApprover={userApprover}
+          selectBox_UserApprover={selectBox_UserApprover}
+          PostAllDatas={PostAllDatas}
+        />
 
-          <div
-            className="mr-8 h-[200px] w-[200px] rounded-lg border-[2px] flex items-center justify-center"
-          >
-            {realstate.imagemImovel && realstate.imagemImovel.length > 0 ? (
-              <img
-                src={realstate.imagemImovel[currentImageIndex]}
-              />
-            ) : (
-              <HouseLine size={50} />
-            )}
-          </div>
-
-          <div className="flex flex-col w-1/3 gap-y-3">
-            <h1 className="text-lg text-gray-700">Imóvel:</h1>
-            <Select
-              value={selectBox_Realstate.selectedOption}
-              onChange={selectBox_Realstate.handleChange}
-              onInputChange={selectBox_Realstate.delayedSearch}
-              loadOptions={selectBox_Realstate.loadOptions}
-              options={selectBox_Realstate.options}
-              placeholder="Pesquisar inscrição cadastral . . ."
-              isClearable
-              isSearchable
-              noOptionsMessage={() => {
-                if (list_Realstate.list.length === 0) {
-                  return "Nenhuma Inscrição Cadastral existente!";
-                } else {
-                  return "Nenhuma opção encontrada!";
-                }
-              }}
-              className="style-select"
-              required
-            />
-
-            <h1 className="text-lg text-gray-700">Número:</h1>
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={realstate.numeroImovel || ''}
-            />
-
-            <h1 className="text-lg text-gray-700">Proprietário:</h1>
-            <div>
-              {owner.imagemPessoa ? (
-                <img
-                  src={owner.imagemPessoa ? owner.imagemPessoa : ""}
-                  className="cursor-pointer rounded-full w-[50px] h-[50px] object-cover p-1 shadow-md"
-                />
-              ) : (
-                <User size={50} />
-              )}
-              <input
-                type="text"
-                disabled
-                className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-                value={owner.nomePessoa || ''}
-              />
-            </div>
-
-            <h1 className="text-lg text-gray-700">Contribuinte:</h1>
-            <div>
-              {taxpayer.imagemPessoa ? (
-                <img
-                  src={taxpayer.imagemPessoa ? taxpayer.imagemPessoa : ""}
-                  className="cursor-pointer rounded-full w-[50px] h-[50px] object-cover p-1 shadow-md"
-                />
-              ) : (
-                <User size={50} />
-              )}
-              <input
-                type="text"
-                disabled
-                className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-                value={taxpayer.nomePessoa || ''}
-              />
-            </div>
-
-            <h1 className="text-lg text-gray-700">Uso:</h1>
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={use.nomeUso || ''}
-            />
-
-            <h1 className="text-lg text-gray-700">Ocupação Atual:</h1>
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={occupation.nomeOcupacaoAtual || ''}
-            />
-          </div>
-
-
-          {/* Tipo Processo: ----------------------------------------------------------------------------------------------------*/}
-
-          <div className="flex flex-col w-1/3 gap-y-3">
-            <h1 className="text-lg text-gray-700">Tipo Processo:</h1>
-            <Select
-              value={selectBox_TypesProcess.selectedOption}
-              onChange={selectBox_TypesProcess.handleChange}
-              onInputChange={selectBox_TypesProcess.delayedSearch}
-              loadOptions={selectBox_TypesProcess.loadOptions}
-              options={selectBox_TypesProcess.options}
-              placeholder="Pesquisar tipo processo . . ."
-              isClearable
-              isSearchable
-              noOptionsMessage={() => {
-                if (list_TypesProcess.list.length === 0) {
-                  return "Nenhuma Tipo Processo existente!";
-                } else {
-                  return "Nenhuma opção encontrada!";
-                }
-              }}
-              className="style-select"
-              required
-            />
-
-            <h1 className="text-lg text-gray-700">Descrição:</h1>
-            <textarea
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb] w-full h-32 resize-none"
-              value={typeProcess.descricaoTipoProcesso || ''}
-            ></textarea>
-
-
-            {/* Processo: ----------------------------------------------------------------------------------------------------*/}
-
-            <h1 className="text-lg text-gray-700">Número de Identificação:</h1>
-            <input
-              type="text"
-              className="rounded-sm border-[#e5e7eb]"
-              onChange={(e) =>
-                setProcess((prevState) => ({
-                  ...prevState,
-                  identificationNumber: e.target.value,
-                }))
-              }
-              value={process.identificationNumber}
-              required
-            />
-
-            <h1 className="text-lg text-gray-700">Situação:</h1>
-            <textarea
-              className="rounded-sm border-[#e5e7eb] w-full h-32 resize-none"
-              onChange={(e) =>
-                setProcess((prevState) => ({
-                  ...prevState,
-                  processSituation: e.target.value,
-                }))
-              }
-              value={process.processSituation}
-            />
-
-            <h1 className="text-lg text-gray-700">Descrição:</h1>
-            <textarea
-              className="rounded-sm border-[#e5e7eb] w-full h-32 resize-none"
-              onChange={(e) =>
-                setProcess((prevState) => ({
-                  ...prevState,
-                  processDescription: e.target.value,
-                }))
-              }
-              value={process.processDescription}
-            />
-
-            <h1 className="text-lg text-gray-700">Data de Aprovação:</h1>
-            <input
-              type="date"
-              className="rounded-sm border-[#e5e7eb]"
-              onChange={(e) =>
-                setProcess((prevState) => ({
-                  ...prevState,
-                  approvationDate: e.target.value,
-                }))
-              }
-              value={process.approvationDate}
-            />
-
-            <h1 className="text-lg text-gray-700">Status:</h1>
-            <input
-              disabled
-              type="text"
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={process.processStatus}
-            />
-
-            <h1 className="text-lg text-gray-700">Engenheiro:</h1>
-            {engineer.imagemPessoa ? (
-              <img
-                className="h-[50px] w-[50px]"
-                src={engineer.imagemPessoa}
-              />
-            ) : (
-              <User size={50} />
-            )}
-            <Select
-              value={selectBox_Engineer.selectedOption}
-              onChange={selectBox_Engineer.handleChange}
-              onInputChange={selectBox_Engineer.delayedSearch}
-              loadOptions={selectBox_Engineer.loadOptions}
-              options={selectBox_Engineer.options}
-              placeholder="Pesquisar engenheiro . . ."
-              isClearable
-              isSearchable
-              noOptionsMessage={() => {
-                if (list_Users.list.length === 0) {
-                  return "Nenhum Engenheiro existente!";
-                } else {
-                  return "Nenhuma opção encontrada!";
-                }
-              }}
-              className="style-select"
-            />
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={engineer.creaEngenheiro || ''}
-            />
-
-            <h1 className="text-lg text-gray-700">Fiscal:</h1>
-            {supervisor.imagemPessoa ? (
-              <img
-                className="h-[50px] w-[50px]"
-                src={supervisor.imagemPessoa}
-              />
-            ) : (
-              <User size={50} />
-            )}
-            <Select
-              value={selectBox_Supervisor.selectedOption}
-              onChange={selectBox_Supervisor.handleChange}
-              onInputChange={selectBox_Supervisor.delayedSearch}
-              loadOptions={selectBox_Supervisor.loadOptions}
-              options={selectBox_Supervisor.options}
-              placeholder="Pesquisar fiscal . . ."
-              isClearable
-              isSearchable
-              noOptionsMessage={() => {
-                if (list_Users.list.length === 0) {
-                  return "Nenhum Fiscal existente!";
-                } else {
-                  return "Nenhuma opção encontrada!";
-                }
-              }}
-              className="style-select"
-            />
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={supervisor.cpfCnpjPessoa || ''}
-            />
-
-
-            {/* Funcionário Responsável */}
-            <h1 className="text-lg text-gray-700">Responsável:</h1>
-            {userResponsible.imagemPessoa ? (
-              <img
-                className="h-[50px] w-[50px]"
-                src={userResponsible.imagemPessoa}
-              />
-            ) : (
-              <User size={50} />
-            )}
-            <Select
-              value={selectBox_UserResponsible.selectedOption}
-              onChange={selectBox_UserResponsible.handleChange}
-              onInputChange={selectBox_UserResponsible.delayedSearch}
-              loadOptions={selectBox_UserResponsible.loadOptions}
-              options={selectBox_UserResponsible.options}
-              placeholder="Pesquisar usuário . . ."
-              isClearable
-              isSearchable
-              noOptionsMessage={() => {
-                if (list_Users.list.length === 0) {
-                  return "Nenhum Usuário existente!";
-                } else {
-                  return "Nenhuma opção encontrada!";
-                }
-              }}
-              className="style-select"
-            />
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={userResponsible.emailPessoa || ''}
-            />
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={typeResponsible.nomeTipoUsuario || ''}
-            />
-
-            <h1 className="text-lg text-gray-700">Aprovador:</h1>
-            {userApprover.imagemPessoa ? (
-              <img
-                className="h-[50px] w-[50px]"
-                src={userApprover.imagemPessoa}
-              />
-            ) : (
-              <User size={50} />
-            )}
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={userApprover.nomePessoa || ''}
-            />
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={userApprover.emailPessoa || ''}
-            />
-            <input
-              type="text"
-              disabled
-              className="cursor-not-allowed rounded-sm border-[#e5e7eb]"
-              value={typeApprover.nomeTipoUsuario || ''}
-            />
-          </div>
-
-          <button className={`btn bg-[#2AA646] text-white hover:text-white hover:bg-[#059669]`} style={{ width: '100px', height: '40px' }} onClick={() => PostAllDatas()} >
-            {'Cadastrar'}
-          </button>
-        </div>
         <hr className="my-10" />
-
-
 
         <DocumentComponent
           stages={stages}
